requirements.txt
absl-py==1.1.0
asgiref==3.5.0
astunparse==1.6.3
boto3==1.21.29
botocore==1.24.29
cachetools==5.2.0
certifi==2020.6.20
charset-normalizer==2.0.12
Django==3.1.4
django-storages==1.12.3
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
jmespath==1.0.0
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
mongoengine==0.24.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.2.4
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.2.0
tzdata==2021.5
urllib3==1.26.9
vaderSentiment==3.3.2
Werkzeug==2.1.2
whitenoise==6.1.0
wincertstore==0.2
wrapt==1.14.1